mac-server/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
websockets>=12.0
python-dotenv>=1.0.0
# torch, diffusers, accelerate, peft, pillow, opencv-python-headless
# already installed in the project venv — no need to reinstall
